+++ torch/csrc/jit/codegen/cuda/scheduler/all_schedulers.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <torch/csrc/jit/codegen/cuda/scheduler/normalization.h>
+#include <torch/csrc/jit/codegen/cuda/scheduler/pointwise.h>
+#include <torch/csrc/jit/codegen/cuda/scheduler/reduction.h>
+#include <torch/csrc/jit/codegen/cuda/scheduler/reduction_heuristic.h>
+
+namespace torch {
+namespace jit {
+namespace fuser {
+namespace cuda {
+
+enum class TORCH_CUDA_API ScheduleHeuristic {
+  PointWise,
+  Reduction,
+  Normalization
+};
+
+}
+} // namespace fuser
+} // namespace jit
+} // namespace torch
