+++ torch/csrc/jit/codegen/cuda/scheduler/pointwise.h
@@ -1,0 +1,22 @@
+#pragma once
+
+#include <ATen/core/ivalue.h>
+
+#include <torch/csrc/jit/codegen/cuda/fusion.h>
+
+namespace torch {
+namespace jit {
+namespace fuser {
+namespace cuda {
+
+// return true or false on whether given fusion could be scheduled;
+TORCH_CUDA_CU_API bool scheduleFusion(
+    Fusion* fusion,
+    const at::ArrayRef<c10::IValue> inputs);
+
+TORCH_CUDA_CU_API bool scheduleFusion(Fusion* fusion);
+
+} // namespace cuda
+} // namespace fuser
+} // namespace jit
+} // namespace torch
