+++ torch/csrc/jit/codegen/cuda/scheduler/reduction_heuristic.h
@@ -1,29 +1,11 @@
 #pragma once
 
-#include <ATen/core/ivalue.h>
 #include <torch/csrc/jit/codegen/cuda/executor_launch_params.h>
-#include <torch/csrc/jit/codegen/cuda/fusion.h>
-#include <torch/csrc/jit/codegen/cuda/ir_all_nodes.h>
 
 namespace torch {
 namespace jit {
 namespace fuser {
 namespace cuda {
-
-enum class TORCH_CUDA_API ScheduleHeuristic {
-  PointWise,
-  Reduction,
-  Normalization
-};
-
-class ExpressionEvaluator;
-
-// return true or false on whether given fusion could be scheduled;
-TORCH_CUDA_CU_API bool scheduleFusion(
-    Fusion* fusion,
-    const at::ArrayRef<c10::IValue> inputs);
-
-TORCH_CUDA_CU_API bool scheduleFusion(Fusion* fusion);
 
 // Parameters the Reduction Heuristic Generates to describe the optimial
 // schedule. Warning: equal operator is intended for use in caching the kernel
@@ -77,38 +59,6 @@ class ReductionParamsHash {
     return attr_hash;
   }
 };
-
-TORCH_CUDA_CU_API c10::optional<ReductionParams> getReductionHeuristics(
-    Fusion* fusion,
-    const at::ArrayRef<c10::IValue>& fusion_inputs,
-    TensorView* red_tv);
-
-TORCH_CUDA_CU_API c10::optional<ReductionParams> getReductionHeuristics(
-    Fusion* fusion,
-    ExpressionEvaluator& evaluator,
-    TensorView* red_tv);
-
-TORCH_CUDA_CU_API void scheduleReduction(
-    Fusion* fusion,
-    const ReductionParams& rparams,
-    TensorView* red_tv,
-    std::vector<TensorView*> outs_of_red);
-
-TORCH_CUDA_API c10::optional<ReductionParams> getNormalizationHeuristics(
-    Fusion* fusion,
-    const at::ArrayRef<c10::IValue>& fusion_inputs,
-    const std::vector<TensorView*>& reduction_tv);
-
-TORCH_CUDA_API c10::optional<ReductionParams> getNormalizationHeuristics(
-    Fusion* fusion,
-    ExpressionEvaluator& evaluator,
-    const std::vector<TensorView*>& reduction_tv);
-
-TORCH_CUDA_API void scheduleNormalization(
-    Fusion* fusion,
-    const ReductionParams& rparams,
-    const std::vector<TensorView*>& reduction_tv,
-    std::vector<TensorView*>& other_tv);
 
 } // namespace cuda
 } // namespace fuser
