+++ torch/csrc/jit/codegen/cuda/scheduler/registry.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include <torch/csrc/jit/codegen/cuda/fusion.h>
-#include <torch/csrc/jit/codegen/cuda/scheduler/scheduler.h>
+#include <torch/csrc/jit/codegen/cuda/scheduler/all_schedulers.h>
 
 namespace torch {
 namespace jit {
